Rename HeaderSearchProvider to HeaderSearchContext for uniformity
