+++ src/app/components/Header/HeaderSearch.tsx
@@ -10,7 +10,7 @@ import SEARCH_PAGE_CONFIG from '../../pages/SearchPage/config'
 import { SearchType } from '../../pages/SearchPage/constants'
 import { queryParam } from '../../pages/SearchPage/query-params'
 import toSearchParams from '../../utils/toSearchParams'
-import { useHeaderSearch } from './HeaderSearchProvider'
+import { useHeaderSearch } from './HeaderSearchContext'
 import useTraverseList from '../../utils/useTraverseList'
 import type { AutoSuggestSearchResult } from './services/auto-suggest/auto-suggest'
 import autoSuggestSearch, { MIN_QUERY_LENGTH } from './services/auto-suggest/auto-suggest'
