+++ src/app/components/Header/auto-suggest/AutoSuggestItem.tsx
@@ -33,7 +33,7 @@ import getDetailPageData from '../../../utils/getDetailPageData'
 import type { AutoSuggestSearchContent } from '../services/auto-suggest/auto-suggest'
 import useParam from '../../../utils/useParam'
 import { routing } from '../../../routes'
-import { useHeaderSearch } from '../HeaderSearchProvider'
+import { useHeaderSearch } from '../HeaderSearchContext'
 
 export interface AutoSuggestItemProps {
   content: string
